+++ Recorder/src/main/java/org/github/ezauton/ezauton/recorder/ISubRecording.java
@@ -1,4 +1,4 @@
-package com.team2502.ezauton.recorder;
+package org.github.ezauton.ezauton.recorder;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import javafx.animation.KeyFrame;
