+++ Recorder/src/main/java/org/github/ezauton/ezauton/recorder/SequentialDataFrame.java
@@ -1,11 +1,11 @@
-package com.team2502.ezauton.recorder;
+package org.github.ezauton.ezauton.recorder;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  * Describes a recording frame about time-based data
  * <br>
- * Essentially a recording for one instaneous second
+ * Essentially a recording for one instant
  */
 public abstract class SequentialDataFrame
 {
@@ -15,4 +15,9 @@ public abstract class SequentialDataFrame
     protected SequentialDataFrame(double time) {this.time = time;}
 
     protected SequentialDataFrame() {}
+
+    public double getTime()
+    {
+        return time;
+    }
 }
