+++ Recorder/src/main/java/org/github/ezauton/ezauton/recorder/base/RobotState.java
@@ -1,6 +1,7 @@
-package com.team2502.ezauton.recorder;
+package org.github.ezauton.ezauton.recorder.base;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import org.github.ezauton.ezauton.recorder.SequentialDataFrame;
 import org.github.ezauton.ezauton.trajectory.geometry.ImmutableVector;
 
 import java.io.Serializable;
@@ -42,11 +43,6 @@ public class RobotState extends SequentialDataFrame implements Serializable
     }
 
     protected RobotState() {}
-
-    public double getTime()
-    {
-        return time;
-    }
 
     public ImmutableVector getPos()
     {
