+++ Visualizer/src/main/java/org/github/ezauton/ezauton/visualizer/IDataProcessor.java
@@ -1,4 +1,4 @@
-package com.team2502.guitools.ppsimulator;
+package org.github.ezauton.ezauton.visualizer;
 
 import javafx.animation.KeyFrame;
